backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0,<11.0
python-dotenv>=1.0,<2.0
gunicorn>=21.0,<23.0
django-filter>=23.0,<25.0
django-allauth>=0.63.0
dj-rest-auth[with_social]>=6.0.0
